refactor piwik auto-creation into an implementation of IDiscussionCreationCallback
