+++ development.ini
@@ -377,6 +377,11 @@ web_analytics_piwik_url =
 # For more information, see http://developer.piwik.org/api-reference/reporting-api#authenticate-to-the-api-via-token_auth-parameter
 web_analytics_piwik_api_token = 
 
+# When a discussion is created, those callbacks will be invoked
+discussion_callbacks =
+    assembl.lib.discussion_creation.DiscussionCreationPrinter
+#    assembl.tasks.piwik.AutomaticPiwikBindingAtDiscussionCreation
+
 [pshell]
 db = assembl.lib.pshell_session.db
 models = assembl.models
